requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
Django==4.2.5
django-cors-headers==4.4.0
djangorestframework==3.15.2
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.34.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.0
h5py==3.11.0
idna==3.8
Jinja2==3.1.4
keras==2.14.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.2
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.3
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.4
psutil==6.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyMySQL==1.1.1
pyparsing==3.0.9
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sqlparse==0.5.1
sympy==1.13.2
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.14.0
termcolor==2.4.0
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.83
ultralytics-thop==2.0.5
urllib3==2.2.2
Werkzeug==3.0.4
wrapt==1.14.1
